backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
groq
sqlalchemy[asyncio]
asyncpg
pytest
python-multipart

